+++ Assets/Scripts/lineScript.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public class lineScript : MonoBehaviour
 {
-    public int colorInt;
+    public double colorInt;
     public Color color;
 }
